requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
pandas
tqdm
sentence-transformers
pyisbn